requirements.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
beautifulsoup4==4.6.0
click==6.7
Cython==0.28.3
Django==2.0.6
djangorestframework==3.8.2
html5lib==0.999999999
macholib==1.5.1
modulegraph==0.10.4
mysql-connector==2.1.6
mysqlclient==1.3.12
nose==1.3.7
numpy==1.14.5
py2app==0.7.3
PyMySQL==0.8.1
pyobjc-core==4.2.2
pyobjc-framework-Cocoa==4.2.2
pyobjc-framework-CoreData==4.2.2
pyobjc-framework-Quartz==4.2.2
pyparsing==2.0.1
pystan==2.17.1.0
python-dateutil==1.5
pytz==2013.7
selenium==3.7.0
six==1.4.1
tornado==5.0.2
virtualenv==16.0.0
webencodings==0.5.1
zope.interface==4.1.1
